Take operator precedence into account
This patch adds the support for operator precedence in ANTLR grammar
and in the generated parser. Previously, binary expressions have been
processed without notion of the operator precedence at all. It lead to
some artifacts in expression formatting with the column limit, when
closely related high-precedence operators may have been broken into two
lines.

The difference can be seen when formatting with comun_limit.
Here is an example having `column_limit = 24`
```
-- Unformatted string
--                     | 24-th column, exceed the limit
local a = a > b and b > c

-- Before:
local a = a > b and b >
        c

-- After:
local a = a > b and
        b > c
```

The patch updates ANTLR to 4.13.2 and migrates to std::any types instead
of ANTLR specific ones.